fixup! Fix serving files that clash with directories
Try and make things more reasable by seperating assignment from if
conditions and removing early returns.

I still kept assignment in the while condition, because it is a nice and
simple way to assign the first non-"/" string popped off from the path
array.

You should be able to autosquash out this commit, since it is a fixup
commit.